fixed bug that prevented pageload for new users
On page load, the app checks if the user has gotten today's current word correct yet by pulling their record from local storage. However, if the user does not have a record in local storage, userStats.date was undefined. This caused errors when trying to get the record's length (to see the last record) and retrieve the date key from an undefined object. I added a ternery to check if the record had a length. If true, it works as normal. If false, it sets an empty new Date('') which is an invalid date but still truthy, allowing the dateIsToday function to still run.